Rename EditReservaation -> ReservationEdit (refs #171)
